change structure of project
