--- guidalsrep/src/main/java/com/guidals/org/data/guidalsrep/GuidalsrepApplication.java
@@ -1,12 +1,0 @@
-package com.guidals.org.data.guidalsrep;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GuidalsrepApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GuidalsrepApplication.class, args);
-	}
-}
