--- guidalsrep/src/main/java/com/guidals/org/data/guidalsrep/books/BookRepo.java
@@ -1,7 +1,0 @@
-package com.guidals.org.data.guidalsrep.books;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface BookRepo extends CrudRepository<Book,Integer>{
-
-}
