--- guidalsrep/src/main/java/com/guidals/org/data/guidalsrep/compsit/CompRepo.java
@@ -1,7 +1,0 @@
-package com.guidals.org.data.guidalsrep.compsit;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface CompRepo extends CrudRepository<Compsit,Integer>{
-
-}
