--- guidalsrep/src/main/java/com/guidals/org/data/guidalsrep/free/FreeRepo.java
@@ -1,7 +1,0 @@
-package com.guidals.org.data.guidalsrep.free;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface FreeRepo extends CrudRepository<Free,Integer>{
-
-}
