--- guidalsrep/src/main/java/com/guidals/org/data/guidalsrep/freesit/FreeSiteRepo.java
@@ -1,7 +1,0 @@
-package com.guidals.org.data.guidalsrep.freesit;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface FreeSiteRepo extends CrudRepository<Freeysit,Integer>{
-
-}
