--- guidalsrep/src/main/java/com/guidals/org/data/guidalsrep/helpfor/HelpforRepo.java
@@ -1,7 +1,0 @@
-package com.guidals.org.data.guidalsrep.helpfor;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface HelpforRepo extends CrudRepository<Helpfor,Integer>{
-
-}
