--- guidalsrep/src/main/java/com/guidals/org/data/guidalsrep/paid/PaidRepo.java
@@ -1,7 +1,0 @@
-package com.guidals.org.data.guidalsrep.paid;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface PaidRepo extends CrudRepository<Paid,Integer>{
-
-}
