--- guidalsrep/src/main/java/com/guidals/org/data/guidalsrep/paidsit/PaidSiteRepo.java
@@ -1,7 +1,0 @@
-package com.guidals.org.data.guidalsrep.paidsit;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface PaidSiteRepo extends CrudRepository<Paidsite,Integer>{
-
-}
