--- guidalsrep/src/main/java/com/guidals/org/data/guidalsrep/suggestions/SuggestionsRepo.java
@@ -1,7 +1,0 @@
-package com.guidals.org.data.guidalsrep.suggestions;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface SuggestionsRepo extends CrudRepository<Suggestions,Integer>{
-
-}
