--- guidalsrep/src/test/java/com/guidals/org/data/guidalsrep/GuidalsrepApplicationTests.java
@@ -1,16 +1,0 @@
-package com.guidals.org.data.guidalsrep;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class GuidalsrepApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
