+++ src/main/java/com/guidals/org/guidals/catalog/GuidalsrepApplication.java
@@ -1,0 +1,12 @@
+package com.guidals.org.guidals.catalog;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GuidalsrepApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GuidalsrepApplication.class, args);
+	}
+}
