+++ src/main/java/com/guidals/org/guidals/catalog/books/BookRepo.java
@@ -1,0 +1,7 @@
+package com.guidals.org.guidals.catalog.books;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface BookRepo extends CrudRepository<Book,Integer>{
+
+}
