+++ src/main/java/com/guidals/org/guidals/catalog/compsit/CompRepo.java
@@ -1,0 +1,7 @@
+package com.guidals.org.guidals.catalog.compsit;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface CompRepo extends CrudRepository<Compsit,Integer>{
+
+}
