+++ src/main/java/com/guidals/org/guidals/catalog/free/FreeRepo.java
@@ -1,0 +1,7 @@
+package com.guidals.org.guidals.catalog.free;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface FreeRepo extends CrudRepository<Free,Integer>{
+
+}
