+++ src/main/java/com/guidals/org/guidals/catalog/freesit/FreeSiteRepo.java
@@ -1,0 +1,7 @@
+package com.guidals.org.guidals.catalog.freesit;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface FreeSiteRepo extends CrudRepository<Freeysit,Integer>{
+
+}
