+++ src/main/java/com/guidals/org/guidals/catalog/helpfor/HelpforRepo.java
@@ -1,0 +1,7 @@
+package com.guidals.org.guidals.catalog.helpfor;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface HelpforRepo extends CrudRepository<Helpfor,Integer>{
+
+}
