+++ src/main/java/com/guidals/org/guidals/catalog/paid/PaidRepo.java
@@ -1,0 +1,7 @@
+package com.guidals.org.guidals.catalog.paid;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface PaidRepo extends CrudRepository<Paid,Integer>{
+
+}
