+++ src/main/java/com/guidals/org/guidals/catalog/paidsit/PaidSiteRepo.java
@@ -1,0 +1,7 @@
+package com.guidals.org.guidals.catalog.paidsit;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface PaidSiteRepo extends CrudRepository<Paidsite,Integer>{
+
+}
