+++ src/main/java/com/guidals/org/guidals/catalog/suggestions/SuggestionsRepo.java
@@ -1,0 +1,7 @@
+package com.guidals.org.guidals.catalog.suggestions;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface SuggestionsRepo extends CrudRepository<Suggestions,Integer>{
+
+}
