+++ src/test/java/com/guidals/org/guidals/catalog/GuidalsrepApplicationTests.java
@@ -1,0 +1,16 @@
+package com.guidals.org.guidals.catalog;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class GuidalsrepApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
